System settings, part II
First, I change the domain class Config. Now it stores all values as
string. I implement the asType() method to convert the Config object to
numbers, booleans, strings, etc.

Last, I implement the LDAP settings page.

+++ grails-app/controllers/org/amcworld/springcrm/DunningController.groovy
@@ -71,14 +71,14 @@ class DunningController {
 		}
 
 		ConfigHolder config = ConfigHolder.instance
-		Integer serviceId = config['serviceIdDunningCharge']
+		Integer serviceId = config['serviceIdDunningCharge'] as Integer
 		if (serviceId) {
 			def service = Service.get(serviceId)
 			if (service) {
 				dunningInstance.addToItems(serviceToItem(service))
 			}
 		}
-		serviceId = config['serviceIdDefaultInterest']
+		serviceId = config['serviceIdDefaultInterest'] as Integer
 		if (serviceId) {
 			def service = Service.get(serviceId)
 			if (service) {
